+++ jax/ops/scatter.py
@@ -16,7 +16,7 @@
 
 
 from .. import lax
-from ..numpy import lax_numpy as jnp
+from jax._src.numpy import lax_numpy as jnp
 
 
 def _scatter_update(x, idx, y, scatter_op):
