+++ jax/scipy/special.py
@@ -20,10 +20,10 @@ import scipy.special as osp_special
 from .. import util
 from .. import lax
 from .. import api
-from ..numpy import lax_numpy as jnp
-from ..numpy.lax_numpy import (asarray, _reduction_dims, _constant_like,
+from jax._src.numpy import lax_numpy as jnp
+from jax._src.numpy.lax_numpy import (asarray, _reduction_dims, _constant_like,
                                _promote_args_inexact)
-from ..numpy._util import _wraps                        
+from jax._src.numpy._util import _wraps
 
 
 @_wraps(osp_special.gammaln)
