+++ jax/scipy/stats/beta.py
@@ -15,9 +15,9 @@
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import (_promote_args_inexact, _constant_like,
-                                where, inf, logical_or)
+from jax._src.numpy._util import _wraps
+from jax._src.numpy.lax_numpy import (_promote_args_inexact, _constant_like,
+                                      where, inf, logical_or)
 from ..special import betaln
