+++ jax/scipy/stats/expon.py
@@ -15,8 +15,8 @@
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact, where, inf
+from jax._src.numpy._util import _wraps
+from jax._src.numpy.lax_numpy import _promote_args_inexact, where, inf
 
 
 @_wraps(osp_stats.expon.logpdf, update_doc=False)
