+++ jax/scipy/stats/logistic.py
@@ -16,8 +16,8 @@ import scipy.stats as osp_stats
 from jax.scipy.special import expit, logit
 
 from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact
+from jax._src.numpy._util import _wraps
+from jax._src.numpy.lax_numpy import _promote_args_inexact
 
 
 @_wraps(osp_stats.logistic.logpdf, update_doc=False)
