+++ jax/scipy/stats/multivariate_normal.py
@@ -19,8 +19,8 @@ import scipy.stats as osp_stats
 from ... import lax
 from ...lax_linalg import cholesky, triangular_solve
 from ... import numpy as jnp
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_dtypes_inexact, _constant_like
+from jax._src.numpy._util import _wraps
+from jax._src.numpy.lax_numpy import _promote_dtypes_inexact, _constant_like
 
 
 @_wraps(osp_stats.multivariate_normal.logpdf, update_doc=False)
