+++ jax/scipy/stats/norm.py
@@ -17,9 +17,9 @@ import numpy as np
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy import lax_numpy as jnp
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like
+from jax._src.numpy import lax_numpy as jnp
+from jax._src.numpy._util import _wraps
+from jax._src.numpy.lax_numpy import _promote_args_inexact, _constant_like
 from .. import special
 
 @_wraps(osp_stats.norm.logpdf, update_doc=False)
