+++ jax/scipy/stats/pareto.py
@@ -16,8 +16,8 @@
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like, inf, where
+from jax._src.numpy._util import _wraps
+from jax._src.numpy.lax_numpy import _promote_args_inexact, _constant_like, inf, where
 
 
 @_wraps(osp_stats.pareto.logpdf, update_doc=False)
