+++ jax/scipy/stats/uniform.py
@@ -16,9 +16,9 @@
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import (_constant_like, _promote_args_inexact,
-                                where, inf, logical_or)
+from jax._src.numpy._util import _wraps
+from jax._src.numpy.lax_numpy import (_constant_like, _promote_args_inexact,
+                                      where, inf, logical_or)
 
 
 @_wraps(osp_stats.uniform.logpdf, update_doc=False)
