+++ tests/host_callback_test.py
@@ -46,6 +46,8 @@ def skip_if_jit_not_enabled():
   if os.getenv("JAX_ENABLE_JIT_PRINT", "false") == "false":
     raise SkipTest("print jit not enabled yet; use JAX_ENABLE_JIT_PRINT env.")
 
+supported_dtypes = sorted(jtu.supported_dtypes(), key=lambda t: t.__name__)
+
 class _TestingOutputStream(object):
   """Use as `output_stream` for tests."""
 
@@ -532,7 +534,7 @@ where: 10
               dtype=dtype,
               nr_args=nr_args) for nr_args in [1, 2]
           for shape in [(), (2,), (2, 3), (2, 3, 4)]
-          for dtype in jtu.supported_dtypes()))
+          for dtype in supported_dtypes))
   def test_jit_types(self, nr_args=2, dtype=np.int16, shape=(2,)):
     if dtype in (np.complex64, np.complex128, np.bool_):
       raise SkipTest(f"id_print jit not implemented for {dtype}.")
